CAN dbV4.9

+++ main/CAN/canDecodeAuto.c
@@ -101,8 +101,8 @@ bool BIMDDeviceError = 0;
 bool BIMDGroundConnectionFault = 0;
 bool BIMDInvalidState = 0;
 float RIsolation = 0;
-uint8_t fIMDPWM = 0;
-uint8_t rIMDPWM = 0;
+float fIMDPWM = 0;
+float rIMDPWM = 0;
 float CMD_TargetBrakeCurrent = 0;
 float CMD_TargetSpeed = 0;
 float rAPPs[2] = {0};
@@ -949,8 +949,8 @@ esp_err_t IMDDataRx(CAN_frame_t stFrame)
     BIMDGroundConnectionFault = (bool)(((stFrame.abData[0] >> 2) & 0x1));
     BIMDInvalidState = (bool)(((stFrame.abData[0] >> 1) & 0x1));
     RIsolation = (float)((float)((((uint16_t)((stFrame.abData[1] >> 0) & 0xFF)) << 8) | ((uint16_t)((stFrame.abData[2] >> 0) & 0xFF))) * 200.0f);
-    fIMDPWM = (uint8_t)(((stFrame.abData[3] >> 0) & 0xFF));
-    rIMDPWM = (uint8_t)(((stFrame.abData[4] >> 0) & 0xFF));
+    fIMDPWM = (float)((float)(((stFrame.abData[3] >> 0) & 0xFF)));
+    rIMDPWM = (float)((float)(((stFrame.abData[4] >> 0) & 0xFF)));
     return ESP_OK;
 }
 
@@ -979,8 +979,8 @@ esp_err_t IMDDataTx(twai_node_handle_t stCANBus)
     stFrame.abData[0] |= (uint8_t)(((((uint32_t)BIMDInvalidState & 0x1) >> 0) & 0x1) << 1);
     stFrame.abData[1] |= (uint8_t)(((((uint32_t)(((float)RIsolation) / 200.0f) & 0xFFFF) >> 8) & 0xFF) << 0);
     stFrame.abData[2] |= (uint8_t)(((((uint32_t)(((float)RIsolation) / 200.0f) & 0xFFFF) >> 0) & 0xFF) << 0);
-    stFrame.abData[3] |= (uint8_t)(((((uint32_t)fIMDPWM & 0xFF) >> 0) & 0xFF) << 0);
-    stFrame.abData[4] |= (uint8_t)(((((uint32_t)rIMDPWM & 0xFF) >> 0) & 0xFF) << 0);
+    stFrame.abData[3] |= (uint8_t)(((((uint32_t)((float)fIMDPWM) & 0xFF) >> 0) & 0xFF) << 0);
+    stFrame.abData[4] |= (uint8_t)(((((uint32_t)((float)rIMDPWM) & 0xFF) >> 0) & 0xFF) << 0);
 
     return CAN_transmit(stCANBus, &stFrame);
 }
